Pull error handling out, and respond to requests with unsupported actions
